// frontend/src/common/api/models/TableCertificatesDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * CertificatesManagerApi
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface TableCertificatesDTO
 */
export interface TableCertificatesDTO {
    /**
     * 
     * @type {string}
     * @memberof TableCertificatesDTO
     */
    handle?: string | null;
    /**
     * 
     * @type {string}
     * @memberof TableCertificatesDTO
     */
    supplier?: string | null;
    /**
     * 
     * @type {string}
     * @memberof TableCertificatesDTO
     */
    certificateType?: string | null;
    /**
     * 
     * @type {string}
     * @memberof TableCertificatesDTO
     */
    validFrom?: string | null;
    /**
     * 
     * @type {string}
     * @memberof TableCertificatesDTO
     */
    validTo?: string | null;
}

/**
 * Check if a given object implements the TableCertificatesDTO interface.
 */
export function instanceOfTableCertificatesDTO(value: object): value is TableCertificatesDTO {
    return true;
}

export function TableCertificatesDTOFromJSON(json: any): TableCertificatesDTO {
    return TableCertificatesDTOFromJSONTyped(json, false);
}

export function TableCertificatesDTOFromJSONTyped(json: any, ignoreDiscriminator: boolean): TableCertificatesDTO {
    if (json == null) {
        return json;
    }
    return {
        
        'handle': json['handle'] == null ? undefined : json['handle'],
        'supplier': json['supplier'] == null ? undefined : json['supplier'],
        'certificateType': json['certificateType'] == null ? undefined : json['certificateType'],
        'validFrom': json['validFrom'] == null ? undefined : json['validFrom'],
        'validTo': json['validTo'] == null ? undefined : json['validTo'],
    };
}

export function TableCertificatesDTOToJSON(value?: TableCertificatesDTO | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'handle': value['handle'],
        'supplier': value['supplier'],
        'certificateType': value['certificateType'],
        'validFrom': value['validFrom'],
        'validTo': value['validTo'],
    };
}